Adiciona refatorção separando em metodos o retorno de categoria e receitas

+++ Challenge_Backend_Financas/Entities/Categorias/GastoCategoriasResponse.cs
@@ -3,6 +3,6 @@
     public class GastoCategoriasResponse
     {
         public string NomeCategoria { get; set; }
-        public double GastoTotal { get; set; }
+        public decimal GastoTotal { get; set; }
     }
 }
