+++ Challenge_Backend_Financas/Entities/ResumoResponse.cs
@@ -4,9 +4,9 @@ namespace Challenge_Backend_Financas.Entities
 {
     public class ResumoResponse : Response
     {
-        public double ReceitasTotal { get; set; }
-        public double DespesasTotal { get; set; }
-        public double SaldoFinal { get; set; }
+        public decimal ReceitasTotal { get; set; }
+        public decimal DespesasTotal { get; set; }
+        public decimal SaldoFinal { get; set; }
         public List<GastoCategoriasResponse> GastoPorCategorias { get; set; }
     }
 }
